Make other integration tests part of the test suite
