+++ src/test/java/org/folio/TestUtils.java
@@ -1,7 +1,7 @@
-package org.folio.finc;
+package org.folio;
 
-import static org.folio.finc.ApiTestBase.APPLICATION_JSON;
-import static org.folio.finc.ApiTestBase.CONTENT_TYPE;
+import static org.folio.ApiTestBase.APPLICATION_JSON;
+import static org.folio.ApiTestBase.CONTENT_TYPE;
 import static org.folio.rest.utils.Constants.MODULE_TENANT;
 
 import io.restassured.RestAssured;
@@ -70,14 +70,19 @@ public class TestUtils {
 
   public static void deployRestVerticle()
       throws ExecutionException, InterruptedException, TimeoutException {
+    deployRestVerticle(true);
+  }
+
+  public static void deployRestVerticle(boolean isTesting)
+    throws ExecutionException, InterruptedException, TimeoutException {
     DeploymentOptions options = new DeploymentOptions();
-    options.setConfig(new JsonObject().put("http.port", verticlePort).put("testing", true));
+    options.setConfig(new JsonObject().put("http.port", verticlePort).put("testing", isTesting));
     deploymentId =
-        vertx
-            .deployVerticle(RestVerticle.class.getName(), options)
-            .toCompletionStage()
-            .toCompletableFuture()
-            .get(30, TimeUnit.SECONDS);
+      vertx
+        .deployVerticle(RestVerticle.class.getName(), options)
+        .toCompletionStage()
+        .toCompletableFuture()
+        .get(30, TimeUnit.SECONDS);
   }
 
   public static void undeployRestVerticle() throws ExecutionException, InterruptedException {
