+++ src/test/java/org/folio/finc/config/ConfigEZBCredentialsIT.java
@@ -6,8 +6,8 @@ import io.restassured.http.ContentType;
 import io.vertx.core.json.Json;
 import io.vertx.ext.unit.junit.Timeout;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.TestUtils;
+import org.folio.ApiTestBase;
+import org.folio.TestUtils;
 import org.folio.rest.jaxrs.model.Credential;
 import org.hamcrest.Matchers;
 import org.junit.AfterClass;
