+++ src/test/java/org/folio/finc/config/ConfigFilesIT.java
@@ -8,8 +8,8 @@ import io.restassured.response.Response;
 import io.vertx.ext.unit.junit.Timeout;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import java.util.UUID;
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.TestUtils;
+import org.folio.ApiTestBase;
+import org.folio.TestUtils;
 import org.folio.rest.jaxrs.model.Isil;
 import org.junit.After;
 import org.junit.AfterClass;
