+++ src/test/java/org/folio/finc/config/ConfigMetadataCollectionsIT.java
@@ -9,8 +9,8 @@ import io.vertx.ext.unit.junit.Timeout;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import java.util.Arrays;
 import java.util.UUID;
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.TestUtils;
+import org.folio.ApiTestBase;
+import org.folio.TestUtils;
 import org.folio.rest.jaxrs.model.FincConfigMetadataCollection;
 import org.folio.rest.jaxrs.model.FincConfigMetadataCollection.MetadataAvailable;
 import org.folio.rest.jaxrs.model.FincConfigMetadataCollection.UsageRestricted;
