+++ src/test/java/org/folio/finc/select/FilterHelperIT.java
@@ -14,8 +14,8 @@ import io.vertx.ext.unit.junit.VertxUnitRunner;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.UUID;
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.TestUtils;
+import org.folio.ApiTestBase;
+import org.folio.TestUtils;
 import org.folio.finc.model.File;
 import org.folio.rest.jaxrs.model.FilterFile;
 import org.folio.rest.jaxrs.model.FincSelectFilter;
