+++ src/test/java/org/folio/finc/select/ITSelectMetadataSourcesHelper.java
@@ -1,34 +1,18 @@
 package org.folio.finc.select;
 
-import io.restassured.RestAssured;
-import io.restassured.parsing.Parser;
-import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.Timeout;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
-import io.vertx.ext.web.client.WebClient;
+
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.ITTestSuiteJunit4;
-import org.folio.finc.TenantUtil;
-import org.folio.finc.TestUtils;
-import org.folio.postgres.testing.PostgresTesterContainer;
-import org.folio.rest.RestVerticle;
-import org.folio.rest.client.TenantClient;
+import org.folio.ApiTestBase;
+import org.folio.TenantUtil;
+import org.folio.TestUtils;
 import org.folio.rest.jaxrs.model.Select;
-import org.folio.rest.jaxrs.model.TenantAttributes;
-import org.folio.rest.persist.PostgresClient;
-import org.folio.rest.tools.utils.NetworkUtils;
-import org.folio.rest.utils.Constants;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Rule;
