+++ src/test/java/org/folio/finc/select/IsilsIT.java
@@ -10,8 +10,8 @@ import io.vertx.core.json.Json;
 import io.vertx.ext.unit.junit.Timeout;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import java.util.UUID;
-import org.folio.finc.ApiTestBase;
-import org.folio.finc.TestUtils;
+import org.folio.ApiTestBase;
+import org.folio.TestUtils;
 import org.folio.rest.jaxrs.model.Isil;
 import org.folio.rest.jaxrs.model.Isils;
 import org.junit.After;
